[CLEANUP] Move the fixture classes around
Also make them final.
