+++ Tests/Unit/Hooks/Fixtures/TestingHookInterfaceNotExtendsHook.php
@@ -2,9 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Seminars\Tests\Unit\Fixtures\Hooks;
-
-use OliverKlee\Seminars\Hooks\Interfaces\Hook;
+namespace OliverKlee\Seminars\Tests\Unit\Hooks\Fixtures;
 
 /**
  * Invalid test interface to use with the HookProviderTest.
